# apiGateway/src/apps/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Simple API User Hobby Test",
        "description": "A REST full OpenAPI Specification for user hobby test"
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:3001",
    "basePath": "/api/v1",
    "paths": {
        "info": "User",
        "/users": {
            "post": {
                "summary": "Creates a user",
                "description": "Adds a new user to the user list.",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "The user to create.",
                        "schema": {
                            "required": [
                                "name"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user succesfully created."
                    }
                }
            },
            "get": {
                "summary": "Gets all users",
                "description": "Returns a list containing all users.",
                "responses": {
                    "200": {
                        "description": "A list of users",
                        "schema": {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "hobbies": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "_id": {
                                                    "type": "string"
                                                },
                                                "passionLevel": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "year": {
                                                    "type": "number"
                                                }
                                            }
                                        }
                                    },
                                    "_id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "summary": "Gets a user",
                "description": "Returns a single user for its id.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The user's id",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A user",
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "type": "array",
                            "items": {
                                "properties": {
                                    "hobbies": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "_id": {
                                                    "type": "string"
                                                },
                                                "passionLevel": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "year": {
                                                    "type": "number"
                                                }
                                            }
                                        }
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "_id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete a user",
                "description": "Delete a selected user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The user's id",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {}
                }
            },
            "put": {
                "summary": "Update a user",
                "description": "Update a new user to the user list.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The user's id",
                        "type": "string"
                    },
                    {
                        "name": "user",
                        "in": "body",
                        "description": "The user to update.",
                        "schema": {
                            "required": [
                                "name"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user updated successfully"
                    }
                }
            }
        },
        "/hobbies": {
            "post": {
                "summary": "Creates a user hobby",
                "description": "Adds a new hobby to the hobby list.",
                "parameters": [
                    {
                        "name": "hobby",
                        "in": "body",
                        "description": "The hobby to create.",
                        "schema": {
                            "required": [
                                "name",
                                "passionLevel",
                                "year",
                                "userId"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "passionLevel": {
                                    "type": "string"
                                },
                                "year": {
                                    "type": "integer"
                                },
                                "userId": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "hobby created successfully."
                    }
                }
            },
            "get": {
                "summary": "Gets all hobbies",
                "description": "Returns all hobbies list.",
                "responses": {
                    "200": {
                        "description": "A list of users hobbies",
                        "schema": {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "_id": {
                                        "type": "string"
                                    },
                                    "passionLevel": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "year": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/hobbies/{id}": {
            "get": {
                "summary": "Gets a user hobby",
                "description": "Returns a single user hobby for its id.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The hobby's id",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A hobby",
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "type": "array",
                            "items": {
                                "properties": {
                                    "_id": {
                                        "type": "string"
                                    },
                                    "passionLevel": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "year": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete a user hobby",
                "description": "Delete a selected user hobby",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The hobby's id",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {}
                }
            },
            "put": {
                "summary": "Update a user hobby",
                "description": "Update a  user to the hobby list.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The hobby's id",
                        "type": "string"
                    },
                    {
                        "name": "bobby",
                        "in": "body",
                        "description": "The hobby to update.",
                        "schema": {
                            "required": [
                                "name",
                                "passionLevel",
                                "year"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "passionLevel": {
                                    "type": "string"
                                },
                                "year": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "updated user hobby successfully."
                    }
                }
            }
        }
    }
}